# backend/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "医院在线预约挂号系统 API",
    "description": "根据当前后端控制器、请求 DTO、测试断言与《项目分工与接口映射.md》整理的可导入 Apifox 的 OpenAPI 3.0 JSON。统一响应格式为 { code, message, data }。\n\n已补充 x-apifox-folder 和 operationId，导入后更容易按模块分组并生成接口结构。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发"
    }
  ],
...
  "paths": {
...
    "/api/payments": {
      "post": {
        "tags": [
          "支付"
        ],
        "summary": "创建支付流水",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequest"
              },
              "example": {
                "businessOrderNo": "AP202607170001",
                "businessType": "appointment",
                "amount": 18.0,
                "payType": 2
              },
              "examples": {
                "createPaymentAppointment": {
                  "summary": "为挂号单创建支付流水",
                  "value": {
                    "businessOrderNo": "AP202607170001",
                    "businessType": "appointment",
                    "amount": 18.0,
                    "payType": 2
                  }
                },
                "createPaymentConsult": {
                  "summary": "为咨询单创建支付流水",
                  "value": {
                    "businessOrderNo": "CO202607170001",
                    "businessType": "consult",
                    "amount": 39.9,
                    "payType": 1
                  }
                },
                "createPaymentUnsupportedBusiness": {
                  "summary": "不支持的业务类型",
                  "value": {
                    "businessOrderNo": "AP202607170001",
                    "businessType": "unknown",
                    "amount": 18.0,
                    "payType": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "paymentNo": {
                          "type": "string",
                          "example": "PAY202607170001"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "createPaymentSuccess": {
                    "summary": "创建支付流水成功",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "paymentNo": "PAY202607170001"
                      }
                    }
                  },
                  "createPaymentFail": {
                    "summary": "业务类型不支持",
                    "value": {
                      "code": 400,
                      "message": "业务类型不支持",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "x-apifox-folder": "支付",
        "operationId": "postApiPayments",
        "x-apifox-summary": "创建支付流水"
      }
    },
    "/api/payments/callback": {
      "post": {
        "tags": [
          "支付"
        ],
        "summary": "支付回调",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCallbackRequest"
              },
              "example": {
                "paymentNo": "PAY202607170001",
                "tradeNo": "TRADE-001",
                "payStatus": 1
              },
              "examples": {
                "paymentCallbackSuccess": {
                  "summary": "支付回调成功",
                  "value": {
                    "paymentNo": "PAY202607170001",
                    "tradeNo": "TRADE-001",
                    "payStatus": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "??",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "null",
                      "nullable": true,
                      "example": null
                    }
                  }
                },
                "examples": {
                  "successNull": {
                    "summary": "??",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "支付",
        "operationId": "postApiPaymentsCallback",
        "x-apifox-summary": "支付回调"
      }
    },
    "/api/payments/{businessOrderNo}": {
      "get": {
        "tags": [
          "支付"
        ],
        "summary": "支付流水详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "businessOrderNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDetailResponse"
                },
                "examples": {
                  "paymentDetail": {
                    "summary": "支付流水详情",
                    "value": {
                      "code": 200,
                      "message": "success",
                      "data": {
                        "paymentNo": "PAY202607170001",
                        "businessOrderNo": "AP202607170001",
                        "amount": 18.0,
                        "payType": 2,
                        "payStatus": 1,
                        "payTime": "2026-07-17T10:05:00"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-apifox-folder": "支付",
        "operationId": "getApiPaymentsBusinessOrderNo",
        "x-apifox-summary": "支付流水详情"
      }
    },
...
    "/api/reviews/my": {
      "get": {
        "tags": [
          "评价"
        ],
        "summary": "我的评价列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageReviewResponse"
                }
              }
            }
          }
        },
        "x-apifox-folder": "评价",
        "operationId": "getApiReviewsMy",
        "x-apifox-summary": "我的评价列表"
      }
    },
    "/api/reviews": {
      "post": {
        "tags": [
          "评价"
        ],
        "summary": "提交评价",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              },
              "example": {
                "orderType": 1,
                "orderId": 1,
                "doctorId": 1,
                "content": "服务很好",
                "rating": 5
              },
              "examples": {
                "createReview": {
                  "summary": "提交评价",
                  "value": {
                    "orderType": 1,
                    "orderId": 1,
                    "doctorId": 1,
                    "content": "服务很好",
                    "rating": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "example": 2
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Bearer token-1-123456789",
            "description": "登录后携带的 Bearer Token"
          }
        ],
        "x-apifox-folder": "评价",
        "operationId": "postApiReviews",
        "x-apifox-summary": "提交评价"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10
        }
      }
    },
...
    "schemas": {
...
      "CreatePaymentRequest": {
        "type": "object",
        "required": [
          "businessOrderNo",
          "businessType",
          "amount",
          "payType"
        ],
        "properties": {
          "businessOrderNo": {
            "type": "string"
          },
          "businessType": {
            "type": "string",
            "enum": [
              "appointment",
              "consult"
            ]
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "payType": {
            "type": "integer"
          }
        }
      },
      "PaymentCallbackRequest": {
        "type": "object",
        "required": [
          "paymentNo",
          "tradeNo",
          "payStatus"
        ],
        "properties": {
          "paymentNo": {
            "type": "string"
          },
          "tradeNo": {
            "type": "string"
          },
          "payStatus": {
            "type": "integer"
          }
        }
      },
...
      "CreateReviewRequest": {
        "type": "object",
        "properties": {
          "orderType": {
            "type": "integer"
          },
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "rating": {
            "type": "integer"
          }
        }
      },
...
      "PageReviewResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "records": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "total": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "pageSize": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "PaymentDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "paymentNo": {
                "type": "string"
              },
              "businessOrderNo": {
                "type": "string"
              },
              "amount": {
                "type": "number"
              },
              "payType": {
                "type": "integer"
              },
              "payStatus": {
                "type": "integer"
              },
              "payTime": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}